implemented Phong Model + reflections

+++ src/materials.rs
@@ -9,10 +9,10 @@ pub enum Material {
 #[derive(Debug, PartialEq, Clone, Copy)]
 pub struct PhongModel {
     pub color: LinSrgb,
-    pub k_s: f64,
-    pub k_d: f64,
-    pub k_a: f64,
-    pub alpha: f64,
+    pub k_s: f32,
+    pub k_d: f32,
+    pub k_a: f32,
+    pub alpha: f32,
 }
 
 // TODO:
